requirements.txt
Django>=4.2,<5.0
mysqlclient
Pillow
python-dotenv
